+++ VisualizationAssociativeQueue/Views/Controls/StackPeek.xaml
@@ -3,7 +3,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:vm="clr-namespace:VisualizationAssociativeQueue.ViewModels"
+             xmlns:vm="clr-namespace:VisualizationAssociativeQueue.ViewModels.Controls"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
